fix(liveness): do not re-render if not neccesary
Here, the useCallback hook creates a memoized version of fetchCreateLiveness that only changes if any of the dependencies change. In this case, the only dependency is AlertError. This way, we ensure that fetchCreateLiveness has the same identity across different renders, and we can safely include it in the dependencies array of useEffect without causing unnecessary reruns of the effect.